Remove async tests and move it into original.
